+++ app/src/main/java/com/romandevyatov/bestfinance/ui/adapters/history/IncomeHistoryAdapter.kt
@@ -1,4 +1,4 @@
-package com.romandevyatov.bestfinance.ui.adapters
+package com.romandevyatov.bestfinance.ui.adapters.history
 
 import android.os.Build
 import android.view.LayoutInflater
