+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/analyze/AnalyzeFragment.kt
@@ -8,10 +8,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.romandevyatov.bestfinance.databinding.FragmentAnalyzeBinding
-import com.romandevyatov.bestfinance.db.entities.mediator.AnalyzeParentData
-import com.romandevyatov.bestfinance.db.entities.mediator.ChildData
-import com.romandevyatov.bestfinance.db.entities.mediator.SubParentData
-import com.romandevyatov.bestfinance.ui.adapters.analyze.ExpandableAnalyzeGroupAdapter
+import com.romandevyatov.bestfinance.ui.adapters.analyze.ExpandableGroupAdapter
 import com.romandevyatov.bestfinance.utils.Constants
 import com.romandevyatov.bestfinance.viewmodels.ExpenseGroupViewModel
 import com.romandevyatov.bestfinance.viewmodels.IncomeGroupViewModel
@@ -25,7 +22,7 @@ class AnalyzeFragment : Fragment() {
 
     private val incomeGroupViewModel: IncomeGroupViewModel by viewModels()
     private val expenseGroupViewModel: ExpenseGroupViewModel by viewModels()
-    private lateinit var expandableAnalyzeGroupAdapter: ExpandableAnalyzeGroupAdapter
+    private lateinit var expandableGroupAdapter: ExpandableGroupAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -40,31 +37,31 @@ class AnalyzeFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val mList: ArrayList<AnalyzeParentData> = ArrayList()
+        val mList: ArrayList<ParentData> = ArrayList()
 
         incomeGroupViewModel.allNotArchivedIncomeGroupWithIncomeSubGroupsIncludingIncomeHistoryAndLiveData.observe(viewLifecycleOwner) { incomeGroupWithIncomeSubGroupsIncludingIncomeHistories ->
-            val apd = AnalyzeParentData(
+            val apd = ParentData(
                 analyzeParentTitle = "Incomings",
                 type = Constants.INCOMINGS_PARENT_TYPE,
                 subParentNestedList = incomeGroupWithIncomeSubGroupsIncludingIncomeHistories
             )
 
             mList.add(apd)
-            expandableAnalyzeGroupAdapter = ExpandableAnalyzeGroupAdapter(mList)
-            binding.analyzeGroupRecycler.adapter = expandableAnalyzeGroupAdapter
+            expandableGroupAdapter = ExpandableGroupAdapter(mList)
+            binding.analyzeGroupRecycler.adapter = expandableGroupAdapter
             binding.analyzeGroupRecycler.layoutManager = LinearLayoutManager(requireContext())
         }
 
         expenseGroupViewModel.allExpenseGroupWithExpenseSubGroupsIncludingExpenseHistoryAndLiveData.observe(viewLifecycleOwner) { expenses ->
-            val apdr = AnalyzeParentData(
+            val apdr = ParentData(
                 analyzeParentTitle = "Expenses",
                 type = Constants.EXPENSES_PARENT_TYPE,
                 subParentNestedListExpenses = expenses
             )
 
             mList.add(apdr)
-            expandableAnalyzeGroupAdapter = ExpandableAnalyzeGroupAdapter(mList)
-            binding.analyzeGroupRecycler.adapter = expandableAnalyzeGroupAdapter
+            expandableGroupAdapter = ExpandableGroupAdapter(mList)
+            binding.analyzeGroupRecycler.adapter = expandableGroupAdapter
             binding.analyzeGroupRecycler.layoutManager = LinearLayoutManager(requireContext())
         }
