--- app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/analyze/ChangeIsExpanded.kt
@@ -1,7 +1,0 @@
-package com.romandevyatov.bestfinance.ui.fragments.analyze
-
-import com.romandevyatov.bestfinance.db.entities.mediator.AnalyzeParentData
-
-interface ChangeIsExpanded {
-    fun doExpandUpdate(tempList: MutableList<AnalyzeParentData>)
-}
