+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/history/ExpenseHistoryFragment.kt
@@ -9,7 +9,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.romandevyatov.bestfinance.databinding.FragmentExpenseHistoryBinding
-import com.romandevyatov.bestfinance.ui.adapters.ExpenseHistoryAdapter
+import com.romandevyatov.bestfinance.ui.adapters.history.ExpenseHistoryAdapter
 import com.romandevyatov.bestfinance.viewmodels.ExpenseHistoryViewModel
 import dagger.hilt.android.AndroidEntryPoint
