+++ app/src/main/java/com/romandevyatov/bestfinance/ui/fragments/history/TransferHistoryFragment.kt
@@ -4,12 +4,11 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContentProviderCompat.requireContext
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.romandevyatov.bestfinance.databinding.FragmentTransferHistoryBinding
-import com.romandevyatov.bestfinance.ui.adapters.TransferHistoryAdapter
+import com.romandevyatov.bestfinance.ui.adapters.history.TransferHistoryAdapter
 import com.romandevyatov.bestfinance.viewmodels.TransferHistoryViewModel
 import dagger.hilt.android.AndroidEntryPoint
